requirements.txt
psycopg2-binary
sqlalchemy
pydantic
requests
mcp[cli]==1.10.1
python-dotenv
ipykernel
httpx